chore(THU-796): move compareSemver to shared
